Add a popper based tooltip
This tooltip allows us to more easily render react elements as the
tooltip comments and removes a dependency on jQuery / bootstrap allowing
us to upgrade to a non-vulnerable version of bootstrap.

I've used a portal for the tooltip because the elements it generates cannot
exist inside an SVG canvas, and conditionally rendering means we won't
spam the DOM with tons of tooltip elements for content that cannot be
seen.

Refs SEN-595